+++ src/main/java/net/quantumfusion/dashloader/cache/blockstates/DashBlockStateData.java
@@ -16,9 +16,9 @@ public class DashBlockStateData {
     @Serialize(order = 0)
     @SerializeNullable(path = {1})
     @SerializeNullable(path = {0})
-    public Map<Integer, Integer> blockstates;
+    public Map<Long, Integer> blockstates;
 
-    public DashBlockStateData(@Deserialize("blockstates") Map<Integer, Integer> blockstates) {
+    public DashBlockStateData(@Deserialize("blockstates") Map<Long, Integer> blockstates) {
         this.blockstates = blockstates;
     }
