+++ src/main/java/net/quantumfusion/dashloader/cache/models/DashBuiltinBakedModel.java
@@ -16,13 +16,13 @@ public class DashBuiltinBakedModel implements DashModel {
     @Serialize(order = 1)
     public DashModelOverrideList itemPropertyOverrides;
     @Serialize(order = 2)
-    public Integer spritePointer;
+    public Long spritePointer;
     @Serialize(order = 3)
     public boolean sideLit;
 
     public DashBuiltinBakedModel(@Deserialize("transformation") DashModelTransformation transformation,
                                  @Deserialize("itemPropertyOverrides") DashModelOverrideList itemPropertyOverrides,
-                                 @Deserialize("spritePointer") Integer spritePointer,
+                                 @Deserialize("spritePointer") Long spritePointer,
                                  @Deserialize("sideLit") boolean sideLit) {
         this.transformation = transformation;
         this.itemPropertyOverrides = itemPropertyOverrides;
